+++ changelog.md
@@ -11,6 +11,7 @@ This is what we are currently working on:
 - Added token ID to /select (e.g., `/sel PrPRGwFEiNKewWym`) -- this is mainly done for internal programming, but could be useful in a macro.
 - Various tweaks to fix modifier bucket size/position issues.
 - Updated /if [Test] { } { } to allow embedded backslashes (to execute multiple commands inside {})
+- Fixed Send Modifier bucket buttons on tooltip
 
 ## History